Revert "refactor: 폴더 구조 변경"
This reverts commit 434b9ab74f958d28adb6d88791512dfc5abc2042.
