+++ bubbletea03-my-app/src/App.js
@@ -1,7 +1,7 @@
 import './App.css';
 
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Home from './page/Home';
+import Home from './component/home/Home';
 
 function App() {
   return (
@@ -10,7 +10,11 @@ function App() {
 				<Routes>
 					<Route path="*" element={<div> Not Found </div>} />
 
-					<Route path="/" element={<Home/>} />
+					<Route path="/" element={
+            <div>
+              <Home/>
+            </div>
+          }/>
 					<Route path="/product-page" element={<div> DD </div>} />
 				</Routes>
 			</BrowserRouter>
